Add files via upload
For loop practice: in example one, In this for loop Aruna and Sushil get printed first only once. Anu will get printed 10 times. Example two: In this for loop, Durga will get printed 10 times. The only difference is that the initialization part is typed out of the for loop. Example three: In this for loop hey will get printed 10 times.  Example Four: In this for loop, Pneumonoultramicroscopicsilicovolcanoconiosis will get printed infinity times.